+++ index.html
@@ -13,7 +13,7 @@
 <body>
     <div class="root">
         <div class="arenas arena1">
-            <form class="control">
+            <form class="control" style="display: none;">
                 <div class="inputWrap">
                     <h2>HIT</h2>
                     <div class="ul">
