+++ utils.js
@@ -1,60 +1,13 @@
 'use strict'
 
-import {player1, player2} from "./players.js";
-import {player_number} from './players.js'
-
 export const $chat = document.querySelector('.chat');
 export const date = new Date();
 export const $formFight = document.querySelector('.control');
 export const $fightButton = document.querySelector('.button');
 export const $arenas = document.querySelector('.arenas');
 
-const logs = {
-    start: 'Часы показывали [time], когда [player1] и [player2] бросили вызов друг другу.',
-    end: [
-        'Результат удара [playerWins]: [playerLose] - труп',
-        '[playerLose] погиб от удара бойца [playerWins]',
-        'Результат боя: [playerLose] - жертва, [playerWins] - убийца',
-    ],
-    hit: [
-        '[playerDefence] пытался сконцентрироваться, но [playerKick] разбежавшись раздробил копчиком левое ухо врага.',
-        '[playerDefence] расстроился, как вдруг, неожиданно [playerKick] случайно раздробил грудью грудину противника.',
-        '[playerDefence] зажмурился, а в это время [playerKick], прослезившись, раздробил кулаком пах оппонента.',
-        '[playerDefence] чесал <вырезано цензурой>, и внезапно неустрашимый [playerKick] отчаянно размозжил грудью левый бицепс оппонента.',
-        '[playerDefence] задумался, но внезапно [playerKick] случайно влепил грубый удар копчиком в пояс оппонента.',
-        '[playerDefence] ковырялся в зубах, но [playerKick] проснувшись влепил тяжелый удар пальцем в кадык врага.',
-        '[playerDefence] вспомнил что-то важное, но внезапно [playerKick] зевнув, размозжил открытой ладонью челюсть противника.',
-        '[playerDefence] осмотрелся, и в это время [playerKick] мимоходом раздробил стопой аппендикс соперника.',
-        '[playerDefence] кашлянул, но внезапно [playerKick] показав палец, размозжил пальцем грудь соперника.',
-        '[playerDefence] пытался что-то сказать, а жестокий [playerKick] проснувшись размозжил копчиком левую ногу противника.',
-        '[playerDefence] забылся, как внезапно безумный [playerKick] со скуки, влепил удар коленом в левый бок соперника.',
-        '[playerDefence] поперхнулся, а за это [playerKick] мимоходом раздробил коленом висок врага.',
-        '[playerDefence] расстроился, а в это время наглый [playerKick] пошатнувшись размозжил копчиком губы оппонента.',
-        '[playerDefence] осмотрелся, но внезапно [playerKick] робко размозжил коленом левый глаз противника.',
-        '[playerDefence] осмотрелся, а [playerKick] вломил дробящий удар плечом, пробив блок, куда обычно не бьют оппонента.',
-        '[playerDefence] ковырялся в зубах, как вдруг, неожиданно [playerKick] отчаянно размозжил плечом мышцы пресса оппонента.',
-        '[playerDefence] пришел в себя, и в это время [playerKick] провел разбивающий удар кистью руки, пробив блок, в голень противника.',
-        '[playerDefence] пошатнулся, а в это время [playerKick] хихикая влепил грубый удар открытой ладонью по бедрам врага.',
-    ],
-    defence: [
-        '[playerKick] потерял момент и храбрый [playerDefence] отпрыгнул от удара открытой ладонью в ключицу.',
-        '[playerKick] не контролировал ситуацию, и потому [playerDefence] поставил блок на удар пяткой в правую грудь.',
-        '[playerKick] потерял момент и [playerDefence] поставил блок на удар коленом по селезенке.',
-        '[playerKick] поскользнулся и задумчивый [playerDefence] поставил блок на тычок головой в бровь.',
-        '[playerKick] старался провести удар, но непобедимый [playerDefence] ушел в сторону от удара копчиком прямо в пятку.',
-        '[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.',
-        '[playerKick] не думал о бое, потому расстроенный [playerDefence] отпрыгнул от удара кулаком куда обычно не бьют.',
-        '[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.'
-    ],
-    draw: 'Ничья - это тоже победа!'
-};
 
-const getTime = () => {
-    const hours = (date.getHours() < 10 ? '0': '') + date.getHours();
-    const minutes = (date.getMinutes() < 10 ? '0': '') + date.getMinutes();
-    const time = hours + ':' + minutes;
-    return time;
-};
+
 
 export const createElement = (tag, className) => {
     const $tag = document.createElement(tag);
@@ -65,109 +18,14 @@ export const createElement = (tag, className) => {
     return $tag;
 };
 
-export function generateLogs (type, player1, player2, playerValue){
-    const time = getTime();
-    let text;   
-    let el;
-    const damage = playerValue;
-
-    switch(type){
-        case 'start': 
-            text = logs[type].replace('[time]', time).replace('[player1]', player1.hero_name).replace('[player2]', player2.hero_name);    
-            el = `<p>${text}</p>`;
-            break;
-        case 'end':
-            text = logs[type][getRandom(logs[type].length - 1)].replace('[playerWins]', player1.hero_name).replace('[playerLose]', player2.hero_name);    
-            el = `<p>${time} ${text}</p>`;
-            break;
-        case 'hit':
-            text = logs[type][getRandom(logs[type].length - 1)].replace('[playerDefence]', player2.hero_name).replace('[playerKick]', player1.hero_name);    
-            el = `<p>${time} ${text} ${damage} ${player2.hp}/100 </p>`;
-            break;
-        case 'defence':
-            text = logs[type][getRandom(logs[type].length - 1)].replace('[playerKick]', player1.hero_name).replace('[playerDefence]', player2.hero_name);    
-            el = `<p>${time} ${text} ${damage} ${player2.hp}/100</p>`;
-            break;
-        case 'draw':
-            text = logs[type];    
-            el = `<p>${time} ${text}</p>`;
-            break;
-        default:
-            text = 'Error';
-            el = `<p>${text}</p>`;
-    };
-    $chat.insertAdjacentHTML('afterbegin', el);
+export const getTime = () => {
+    const hours = (date.getHours() < 10 ? '0': '') + date.getHours();
+    const minutes = (date.getMinutes() < 10 ? '0': '') + date.getMinutes();
+    const time = hours + ':' + minutes;
+    return time;
 };
 
-export const showResult = () => {
-    if (player1.hp === 0 || player2.hp === 0){
-        $fightButton.disabled = true;
-        $arenas.appendChild(createReloadButton());
-    };
+export const getRandom = (damage) => Math.ceil(Math.random() * damage);
 
-    if (player1.hp === 0 && player1.hp < player2.hp){
-        $arenas.appendChild(player2.playerWin());
-        generateLogs('end', player2, player1);
-
-    } else if (player2.hp === 0 && player2.hp < player1.hp){
-        $arenas.appendChild(player1.playerWin());
-        generateLogs('end', player1, player2);
-
-    } else if (player1.hp === 0 && player2.hp === 0){
-        $arenas.appendChild(playerWin());
-        generateLogs('draw')
-    };
-};
-
-
-const getRandom =(damage) => Math.ceil(Math.random() * damage);
-
-const createReloadButton = () => {
-    const $reloadWrap = createElement('div', 'reloadWrap')
-    const $buttonReload = createElement('button', 'button');
-    $buttonReload.innerText = 'Restart';
-    $reloadWrap.appendChild($buttonReload);
-    $buttonReload.addEventListener('click', function(){
-        window.location.reload();
-    });
-
-    return $reloadWrap;
-};
-
-const ATTACK = ['head', 'body', 'foot'];
-
-const HIT = {
-    head: 30,
-    body: 25,
-    foot: 20,
-};
-
-export const playerAttack = () => {
-    const attack = {};
-    for (let item of $formFight){
-        if (item.checked && item.name === 'hit'){
-            attack.value = getRandom(HIT[item.value]);
-            attack.hit = item.value;
-        };
-
-        if(item.checked && item.name === 'defence'){
-            attack.defence = item.value;
-        };
-        
-        item.checked = false;
-    }
-
-    return attack;
-};
-
-export const enemyAttack = () => {
-    const hit = ATTACK[getRandom(3) - 1];
-    const defence = ATTACK[getRandom(3) - 1];
-    return{
-        value: getRandom(HIT[hit]),
-        hit,
-        defence,
-    }
-};
